:truck: move setup into new architecture
